==> step5.ipynb <==
import os
import sys
import pandas as pd
import numpy as np

df_vcf=pd.read_csv('../step1/output4_GT.tsv', sep='\t')
# df_vcf=df_vcf.drop(['#CHROM', 'POS', 'REF', 'ALT', 'INFO', 'FORMAT'], axis=1)
df_synonymous=pd.read_csv('../step2/list_synonymous.tsv')
df_synonymous=df_synonymous.drop_duplicates()
df_vcf_synonymous=df_vcf[df_vcf['ID'].isin(df_synonymous['variant'])]
df_synonymous_summary=pd.DataFrame()
df_synonymous_summary['IID']=df_vcf_synonymous.columns[1:]
list_count=[]
for column in df_vcf_synonymous.columns[1:]:
#     print(column)
    column_counts = df_vcf_synonymous[column].tolist().count('0/1')+df_vcf_synonymous[column].tolist().count('1/1')*2
    list_count.append(column_counts)
df_synonymous_summary['count_synonymous']=list_count
display(df_synonymous_summary)

df1=pd.read_csv('/data/Segre_Lab/users/jlama/WES_new.ALL_050824/Phenotype/FAME.all.covariates.532IDs.100924.tsv', sep=' ')
display(df1)
pheno = df1.loc[:,['FID', 'IID', 'responder']]
display(pheno)

df_batch1=pd.read_csv('/data/Segre_Lab/users/jlama/WES_new.ALL_050824/PCA/FAME/study_pca.eigenvec', sep='\t')
df_batch1=df_batch1.loc[:,['IID','PC1','PC2','PC3','PC4','PC5','PC6','PC7','PC8','PC9','PC10']]
display(df_batch1)

df_batch1_merge = pd.merge(df_batch1, df1, on=['IID'], how='left')
df_batch1_merge=df_batch1_merge.loc[:,['FID','IID','Age','Sex','steroid.medication','hx.glaucoma','drops.during6m','PC1','PC2','PC3','PC4','PC5','PC6','PC7','PC8','PC9','PC10']]
display(df_batch1_merge)

#list_age=[]
#for row in df_batch1_merge.itertuples():
 #   tmp_age=getattr(row, 'Sex')
 #   new_age=tmp_age
 #   list_age.append(new_age)
#df_batch1_merge['Sex']=list_age
#display(df_batch1_merge)

#df_batch1_merge=df_batch1_merge.loc[:,['IID','Age','Sex','PC1','PC2','PC3','PC4','PC5','PC6','PC7','PC8','PC9','PC10']]
df_batch1_merge=pd.merge(df_batch1_merge, df_synonymous_summary, on=['IID'], how='left')
display(df_batch1_merge)

df_batch1_merge.to_csv('FAME.GWAS_cov_pca_100924.csv', sep='\t', index=None)

pheno.to_csv("FAME.Responder.pheno.csv", sep='\t', index=None)
