==> step6.ipynb <==
import os
import sys
import pandas as pd
import numpy as np

df=pd.read_csv('IDs_phenotypes_FINAL_102424_113023_T2.csv',sep='\t')
display(df)
df['FID']=df['IID']
df=df.loc[:,['FID','IID','Phenotype']]
display(df)
df.to_csv('IDs_phenotypes_FINAL_102424_113023_T2_FID_IID_adjust.csv', sep='\t', index=None)

df=pd.read_csv('GWAS_cov_pca_batch1_072524.csv',sep='\t')
df1=pd.read_csv('IDs_phenotypes_FINAL_102424_113023_T2_FID_IID_adjust.csv',sep='\t')
df['FID']=df['IID']
df=df.loc[:,['FID','IID','Age','Sex','PC1','PC2','PC3','PC4','PC5','PC6','PC7','PC8','PC9','PC10','count_synonymous']]
display(df)
display(df1)
df2=pd.merge(df, df1, on=['FID','IID'], how='inner')
display(df2)
df2=df2.loc[:,['FID','IID','Age','Sex','PC1','PC2','PC3','PC4','PC5','PC6','PC7','PC8','PC9','PC10','count_synonymous']]
df2.to_csv('GWAS_cov_pca_batch1_072524_FID_IID.csv', sep='\t', index=None)

df=pd.read_csv('list_sample.tsv',sep='\t',header=None)
df.columns=['IID']
display(df)
df['FID']=df['IID']
df=df.loc[:,['FID','IID']]
display(df)
df.to_csv('list_sampleID_Batch1_FID_IID.csv', sep='\t', index=None)
