==> Step7.a.ipynb <==
import os
import sys
import numpy as np
import pandas as pd

df_bim=pd.read_csv('../step1/output4.bim', sep='\t', header=None)
df_bim.columns=['CHR','ID','VAR1','POS','ALT','REF']
display(df_bim)
df_bim=df_bim.loc[:,['ID']]
display(df_bim)

df_bim['MAF']=1e-03
display(df_bim)

df_bim.to_csv('AAF_list.tsv', sep='\t', header=None, index=None)
